Quote rc commands if necessary
Surround an argument with quotes before assembling it into a command
base on whether a command requires quoting to be parsed by rc as it
was received through args. If the arg contains any of rc's special
characters or whitespace as documented in rc(1), it is quoted and
any ' characters are escaped to ''.

The quoted command string is also appended to the dump command.

Fixes #135